Lock mouse cursor to game window
Now this is a bit tricky to exit in the Unity editor. The default
hotkey to exit should be CTRL-P, but that doesn't seem to work on
Unity 2019.2.0f1 in a Ubuntu 18.04.2 system.
However, the menu bar can be activated by pressing F10 here, which
allows to navigate to the Edit menu in order to select "Play" (which
stops playback).